Latest edits

+++ src/libs/decoders/SDL_sound.c
@@ -410,7 +410,7 @@ static int init_sample(const Sound_DecoderFunctions *funcs,
     internal->funcs = funcs;
     if (!funcs->open(sample, ext))
     {
-        SDL_RWseek(internal->rw, pos, SEEK_SET);  /* set for next try... */
+        SDL_RWseek(internal->rw, (int)pos, SEEK_SET);  /* set for next try... */
         return(0);
     } /* if */
